ryugaku_assistant/requirements.txt
pandas
openpyxl
streamlit
google-generativeai
